pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "diarized_transcriber"
version = "0.1.0"
description = "A library for transcribing media content with speaker diarization"
readme = "README.md"
authors = [
    { name = "Eddie Dunn", email = "eddie@eddiedunn.com" },
]
license = { text = "MIT" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
keywords = ["transcription", "diarization", "whisperx", "audio", "speech-to-text"]
dependencies = [
    "pydantic",
    "whisperx",
    "torch",
    "torchaudio",
    "faster-whisper",
    "ctranslate2",
    "transformers",
    "pandas",
    "setuptools",
    "nltk",
    "pyannote.audio",
    "soundfile",
]
requires-python = "==3.10"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "mypy",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/diarized_transcriber"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
select = ["E", "F", "B", "I"]
ignore = []
target-version = "py310"